version 0.103
Version 0.103 does not work. Version 0.103.5 adds jump right for P2 and the game functions as it should (95% of the time) - there is an issue in which a possible eatpiece is not detected. Kings are still not implemented and neither is a way to tell if a player has won the game.